backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Gemini Project Dashboard - Local project management and documentation hub"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "fastapi[standard]>=0.122.0",
    "pillow>=10.0.0",
    "pydantic>=2.12.5",
    "pynput>=1.8.1",
    "pystray>=0.19.0",
    "pywebview>=6.1",
    "pyyaml>=6.0.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]

[project.scripts]
dashboard-tray = "backend.tray:main"
